requirements.txt
aiogram==3.17.0
aiosqlite==0.20.0
apscheduler==3.10.4
pytz==2024.1
python-dotenv==1.0.1
